AVE-3164 - Speed splits aggregations (#80)
* AVE-3032 - Speed Aggregation & AVE-3033 - Split Monitor & AVE-3034 - Split Fail

Adding in all the nessisary stuff for the Aggregation Workflows.

* Update exit code handling in HostedServiceBase.cs

Modify exit codes in try-catch block to reflect success
and failure accurately. Set exit code to 0 for successful
processing and 1 for exceptions, improving application
signal for process outcomes.

* Add DecodeEventsCommand for event log processing

This commit introduces a new `DecodeEventsCommand` in the `EventLogCommands.cs` file, enabling the import and decoding of event logs. A corresponding class `DecodeEventsCommand` is created to handle command options and service registration.

Additionally, a new configuration class `DecodeEventsConfiguration` is added to specify the path for event logs. The `DecodeEventLogHostedService` class is introduced to process the logs, grouping them by device identifiers and sending them to a decoding workflow.

A new workflow class `DecodeEventLogWorkflow` is also added, defining the steps for decoding event logs. License comments have been included in the new files to comply with copyright requirements.

* change the codes to mathc

* Updated watchdog to work with GCP secrets

* Create manual-version-publish.yml

* Update manual-version-publish.yml

* Update manual-version-publish.yml

* Create Test.yml

* Update Test.yml

* Add SortOption and update configuration binding

- Introduced a new command option `SortOption` in `GenerateCommand.cs` to enhance sorting functionality.
- Changed configuration binding from `DeviceEventLoggingConfiguration` to `WatchdogConfiguration` for better alignment with application structure.

* Adding in the updated Aggregations (#78)

* Adding in the updated Aggregations

* pr fixes

* update aggs

* take out unnessisary thing

---------

Co-authored-by: Christian Baker <christianbaker@utah.gov>
Co-authored-by: Christian Baker <86611690+ChesireFreeThrow@users.noreply.github.com>